game_manager/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-extensions
djangorestframework
django-cors-headers